Trabajo ya entregado

+++ Application/Response/MercaderiaGetResponse.cs
@@ -4,7 +4,7 @@
     {
         public int id { get; set; }
         public string nombre { get; set; }
-        public int Precio { get; set; }
+        public double Precio { get; set; }
         public TipoMercaderiaResponse tipo { get; set; }
         public string imagen { get; set; }
     }
